src/modules/cart_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
psycopg2-binary
httpx
pymysql
